requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
ezdxf==1.1.4
shapely==2.0.2
matplotlib==3.8.2
numpy==1.24.4
scipy==1.11.4
networkx==3.2.1
opencv-python==4.8.1.78
PyMuPDF==1.23.14
Pillow==10.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
psutil==5.9.6